requirements.txt
annotated-types==0.7.0
argcomplete==3.6.2
bleak==0.22.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
dbus-fast==2.44.3
dotmap==1.3.30
greenlet==3.2.4
idna==3.10
meshtastic==2.7.1
more-itertools==10.7.0
packaging==24.2
print-color==0.4.6
protobuf==6.32.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyMySQL==1.1.1
Pypubsub==4.0.3
PyQRCode==1.2.1
pyserial==3.5
pytap2==2.3.0
python-dotenv==1.1.1
python-ulid==1.1.0
PyYAML==6.0.2
requests==2.32.5
setuptools==80.9.0
SQLAlchemy==2.0.43
tabulate==0.9.0
types-cffi==1.17.0.20250809
types-pyOpenSSL==24.1.0.20240722
types-setuptools==80.9.0.20250809
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
